postgres_rls/pyproject.toml
[tool.poetry]
name = "postgres-rls"
version = "0.1.0"
description = ""
authors = ["Justin Van Winkle <justin.vanwinkle@gmail.com>"]
license = "GPL3"
readme = "README.md"
packages = [{include = "postgres_rls"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg = "^3.1.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
